Mark data/themes.json as a binary file
This adds a .gitattributes file that marks data/themes.json as binary.
data/themes.json stores all it's data on a single line, and that entire
line
is marked as changed between commits. By marking it as binary, git
displays diffs more accurately.

NB: As an alternative to marking it as binary, we could also modify
tools/generate.sh so that it doesn't generate the entire file on one
line. This can be done by removing the `tr -d " \t\n\r"` portion of the
script. Doing so would ensure that git only stored the sections of the
themes.json file that changed when a theme was modified or added.

Signed-off-by: Alecto Irene Perez <perez.cs@pm.me>