Use a column display for themes.
I wish we could just pipe the theme list to /usr/bin/column, but:
  - column has a hardcoded tab size of 8
  - fails catastrophically when the terminal tab size is not 8
  - uses hard-coded tabs for formatting (no options to use spaces)
  - does not support ANSI escape codes (a fix was merged in Feb, but not released yet)

Since the algorithm is trivial, we might as well implement it in bash.

Note: the columns are as large as the largest entry.
Currently, the biggest entry is solarized-dark-higher-contrast
with a length of 30, so columns are 40 characters wide, inc. ( NUM ).